add dummy cache to tests and implement caching tests
