+++ Tests/API/visual_crossing_api_service_tests.py
@@ -1,8 +1,6 @@
 import unittest
 from datetime import datetime
 from unittest.mock import patch
-
-from django.core.cache import cache
 from django.test import override_settings, TestCase
 from requests.exceptions import HTTPError
 from API.visual_crossing_api_service import get_weather_data, WeatherData
@@ -71,40 +69,6 @@ class VisualCrossingAPIServiceTests(TestCase):
         mock_get.assert_called_once()
         mock_get.json.assert_not_called()
         mock_response.raise_for_status.assert_called_once()
-
-    @patch('requests.get')
-    @patch('django.core.cache.backends.dummy.DummyCache')
-    def test_get_weather_data_assert_cache_miss(self, mock_cache, mock_get):
-        # arrange
-        mock_cache_instance = mock_cache.return_value
-        mock_cache_instance.get.return_value = None
-        mock_response = mock_get.return_value
-        mock_response.status_code = 200
-        mock_response.json.return_value = self.mock_json_data
-
-        # act
-        get_weather_data('London', datetime.now())
-
-        # assert
-        mock_cache_instance.get.assert_called_once()
-        mock_cache_instance.set.assert_called_once()
-
-    @patch('requests.get')
-    @patch('django.core.cache.backends.dummy.DummyCache')
-    def test_get_weather_data_assert_cache_hits(self, mock_cache, mock_get):
-        # arrange
-        mock_cache_instance = mock_cache.return_value
-        mock_cache_instance.get.return_value = {'London_2024-09-21': WeatherData}
-        mock_response = mock_get.return_value
-        mock_response.status_code = 200
-        mock_response.json.return_value = self.mock_json_data
-
-        # act
-        get_weather_data('London', datetime.now())
-
-        # assert
-        mock_cache_instance.get.assert_called_once()
-        mock_cache_instance.set.assert_not_called()
 
     mock_json_data = {
             'resolvedAddress': 'London, United Kingdom',
